This commit was manufactured by cvs2git to create branch 'MIRBSD_8'.

+++ src/sys/ddb/db_command.c
@@ -1,28 +1,29 @@
+/**	$MirOS$ */
 /*	$OpenBSD: db_command.c,v 1.35 2004/04/25 03:21:50 itojun Exp $	*/
 /*	$NetBSD: db_command.c,v 1.20 1996/03/30 22:30:05 christos Exp $	*/
 
-/* 
+/*
  * Mach Operating System
  * Copyright (c) 1993,1992,1991,1990 Carnegie Mellon University
  * All Rights Reserved.
- * 
+ *
  * Permission to use, copy, modify and distribute this software and its
  * documentation is hereby granted, provided that both the copyright
  * notice and this permission notice appear in all copies of the
  * software, derivative works or modified versions, and any portions
  * thereof, and that both notices appear in supporting documentation.
- * 
+ *
  * CARNEGIE MELLON ALLOWS FREE USE OF THIS SOFTWARE IN ITS "AS IS"
  * CONDITION.  CARNEGIE MELLON DISCLAIMS ANY LIABILITY OF ANY KIND FOR
  * ANY DAMAGES WHATSOEVER RESULTING FROM THE USE OF THIS SOFTWARE.
- * 
+ *
  * Carnegie Mellon requests users of this software to return to
- * 
+ *
  *  Software Distribution Coordinator  or  Software.Distribution@CS.CMU.EDU
  *  School of Computer Science
  *  Carnegie Mellon University
  *  Pittsburgh PA 15213-3890
- * 
+ *
  * any improvements or extensions that they make and grant Carnegie Mellon
  * the rights to redistribute these changes.
  */
@@ -295,7 +296,7 @@ db_map_print_cmd(addr, have_addr, count, modif)
 	char *		modif;
 {
         boolean_t full = FALSE;
-        
+
         if (modif[0] == 'f')
                 full = TRUE;
 
@@ -330,7 +331,7 @@ db_object_print_cmd(addr, have_addr, count, modif)
 	char *		modif;
 {
         boolean_t full = FALSE;
-        
+
         if (modif[0] == 'f')
                 full = TRUE;
 
@@ -346,7 +347,7 @@ db_page_print_cmd(addr, have_addr, count, modif)
 	char *		modif;
 {
         boolean_t full = FALSE;
-        
+
         if (modif[0] == 'f')
                 full = TRUE;
 
@@ -601,7 +602,7 @@ db_fncall(addr, have_addr, count, modif)
 
 	retval = (*func)(args[0], args[1], args[2], args[3], args[4],
 			 args[5], args[6], args[7], args[8], args[9]);
-	db_printf("%#n\n", retval);
+	db_printf("%#n\n", (int)retval);
 }
 
 void
